requirements.txt
absl-py==2.1.0
accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
boto3==1.36.15
botocore==1.36.15
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.3
coverage==7.3.2
crispy-bootstrap5==0.7
cryptography==44.0.0
cssselect2==0.7.0
cycler==0.12.1
datasets==3.6.0
debugpy==1.8.9
decorator==5.1.1
deepface==0.0.93
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.9
Django==5.2.4
django-allauth==0.61.0
django-cors-headers==4.7.0
django-crispy-forms==2.1
django-otp==1.3.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.4.0
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
executing==2.1.0
factory-boy==3.3.0
Faker==20.1.0
filelock==3.18.0
fire==0.7.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
flatbuffers==25.1.24
fonttools==4.58.2
frozenlist==1.7.0
fsspec==2025.3.0
gast==0.6.0
gdown==5.2.0
geographiclib==2.1
geopy==2.4.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.179.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.63.0rc1
gunicorn==23.0.0
h5py==3.12.1
httplib2==0.22.0
huggingface-hub==0.33.2
idna==3.10
iniconfig==2.1.0
instaloader==4.14.1
ipykernel==6.29.5
ipython==8.30.0
ipywidgets==8.1.7
itsdangerous==2.2.0
jax==0.5.0
jaxlib==0.5.0
jedi==0.19.2
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.15
keras==3.8.0
kiwisolver==1.4.8
libclang==18.1.1
lxml==6.0.0
lz4==4.4.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.21
ml-dtypes==0.4.1
MouseInfo==0.1.3
mpmath==1.3.0
mtcnn==1.0.0
multidict==6.6.3
multiprocess==0.70.16
mysql-connector-python==9.1.0
mysqlclient==2.2.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
oauthlib==3.2.2
opencv-contrib-python==4.11.0.86
opencv-python==4.8.1.78
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.14.0
packaging==25.0
pandas==2.3.1
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
propcache==0.3.2
proto-plus==1.26.1
protobuf==4.25.6
psutil==6.1.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pybind11==2.13.6
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydyf==0.11.0
PyGetWindow==0.0.9
Pygments==2.19.1
PyJWT==2.10.1
PyMsgBox==1.0.9
PyMySQL==1.1.1
pyparsing==3.2.3
pyperclip==1.9.0
pyphen==0.17.2
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
pyqt5-tools==5.15.9.3.3
PyQt5_sip==12.16.1
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytest==7.4.3
pytest-django==4.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python3-openid==3.2.0
pytweening==1.2.0
pytz==2025.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
regex==2024.11.6
reportlab==4.4.3
requests==2.32.4
requests-oauthlib==2.0.0
retina-face==0.0.17
rich==13.9.4
rsa==4.9
s3transfer==0.11.2
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.15.3
seaborn==0.13.2
sentence-transformers==5.0.0
sentencepiece==0.2.0
six==1.17.0
snscrape==0.7.0.20230622
sounddevice==0.5.1
soupsieve==2.6
SQLAlchemy==2.0.23
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.14.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tf_keras==2.18.0
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.2
torch==2.7.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.28.1
wcwidth==0.2.13
weasyprint==64.0
webencodings==0.5.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
zopfli==0.2.3.post1
